+++ src/main/java/com/example/demo/service/board/BoardFacade.java
@@ -27,6 +27,7 @@ import com.example.demo.service.project.ProjectServiceImpl;
 import com.example.demo.service.project.ProjectTechnologyService;
 import com.example.demo.service.technology_stack.TechnologyStackService;
 import com.example.demo.service.trust_grade.TrustGradeService;
+import com.example.demo.service.trust_grade.TrustGradeServiceImpl;
 import com.example.demo.service.user.UserProjectHistoryService;
 import com.example.demo.service.user.UserService;
 import java.util.ArrayList;
